new coding

+++ DSA In Java/Hashing/StorFre.java
@@ -6,25 +6,26 @@ public class StorFre {
         String s = "successes";
         Map<Character, Integer> freqMap = new HashMap<>();
         for (char c : s.toCharArray()) {
-            if (c == ' ') continue; // skip spaces
+            if (c == ' ')
+                continue; // skip spaces
             freqMap.put(c, freqMap.getOrDefault(c, 0) + 1);
         }
-         int mxv = 0;
-         int mx = 0;
+        int mxv = 0;
+        int mx = 0;
         for (Map.Entry<Character, Integer> e : freqMap.entrySet()) {
 
-            if(e.getKey() == 'a' || e.getKey() == 'e' || e.getKey() == 'i' || e.getKey() == 'o' || e.getKey() == 'u') {
-                if(e.getValue() > mxv) {
-                    mxv = e.getValue();
+            if (e.getKey() == 'a' || e.getKey() == 'e' || e.getKey() == 'i' || e.getKey() == 'o' || e.getKey() == 'u') {
+                if (e.getValue() > mxv) {
+                    mxv = e.getValue().intValue();
                 }
-            }else{
-                if(e.getValue() > mx){
+            } else {
+                if (e.getValue() > mx) {
                     mx = e.getValue();
                 }
             }
         }
-        
-        System.out.println(mx+mxv);
+
+        System.out.println(mx + mxv);
 
     }
 }
